requirements.txt
torch
torchsde
torchvision
torchaudio
einops
transformers>=4.28.1
tokenizers>=0.13.3
sentencepiece
safetensors>=0.4.2
aiohttp
pyyaml
Pillow
scipy
tqdm
psutil
python-dotenv
anthropic

#non essential dependencies:
kornia>=0.7.1
spandrel
soundfile